+++ archsim/modules/arm-system-devices/inc/GenericPrimecellDevice.h
@@ -29,8 +29,8 @@ namespace archsim
 				COMPONENT_PARAMETER_U64(PeripheralID);
 				COMPONENT_PARAMETER_U64(PrimecellID);
 
-				bool Read(uint32_t offset, uint8_t size, uint32_t& data) override;
-				bool Write(uint32_t offset, uint8_t size, uint32_t data) override;
+				bool Read(uint32_t offset, uint8_t size, uint64_t& data) override;
+				bool Write(uint32_t offset, uint8_t size, uint64_t data) override;
 
 				bool Initialise() override;
 			};
