+++ archsim/src/abi/devices/virtio/VirtIO.cpp
@@ -85,7 +85,7 @@ VirtIO::~VirtIO()
 	}
 }
 
-bool VirtIO::Read(uint32_t offset, uint8_t size, uint32_t& data)
+bool VirtIO::Read(uint32_t offset, uint8_t size, uint64_t& data)
 {
 	LC_DEBUG3(LogVirtIO) << "[" << GetName() << "] Register Read offset=" << std::hex << offset << ", size=" << std::dec << (uint32_t)size;
 	if (offset >= 0x100 && offset <= (0x100 + GetConfigAreaSize())) {
